requirements.txt
numpy
scipy
skimage
trimesh
tqdm
matplotlib
pyvista
graph_tool
pynrrd
sklearn